agrampar.y: Remove 'Embedded' nonterminal.
The previous grammar misleadingly implied that every occurrence
of embedded code could be either surrounded with braces or
terminated with a semicolon, but in fact the lexer only allows
one or the other in each circumstance.  The new grammar makes
that clear by substituting the appliable RHS of the Embedded
nonterminal for each occurrence.  Then, since Embedded no longer
appeared in any RHS, I removed the nonterminal.

Regenerate agrampar.tab.cc.